Revert "Merge branch 'programmer-branch' into eric-branch"
This reverts commit 7d2a9f1e63c7f60e437729a62d9cde23029afb50, reversing
changes made to 9283eb82d3902a26d15fb757f0d922d2cd4c71d4.
